Move Gradle/Maven testing commands to test-utils
