+++ test-cli/src/test/groovy/io/micronaut/starter/cli/test/CreateControllerSpec.groovy
@@ -3,23 +3,15 @@ package io.micronaut.starter.cli.test
 import io.micronaut.starter.application.ApplicationType
 import io.micronaut.starter.cli.CodeGenConfig
 import io.micronaut.starter.cli.feature.server.controller.CreateControllerCommand
-import io.micronaut.starter.test.CommandSpec
-import io.micronaut.starter.test.LanguageBuildTestFrameworkCombinations
 import io.micronaut.starter.io.ConsoleOutput
 import io.micronaut.starter.options.BuildTool
 import io.micronaut.starter.options.Language
 import io.micronaut.starter.options.TestFramework
-import org.apache.maven.cli.MavenCli
-import org.gradle.testkit.runner.BuildResult
-import org.gradle.testkit.runner.GradleRunner
-import spock.lang.Shared
+import io.micronaut.starter.test.CommandSpec
+import io.micronaut.starter.test.LanguageBuildTestFrameworkCombinations
 import spock.lang.Unroll
 
-import java.nio.charset.StandardCharsets
-
 class CreateControllerSpec extends CommandSpec {
-    @Shared GradleRunner gradleRunner = GradleRunner.create()
-    @Shared MavenCli cli = new MavenCli()
 
     @Unroll
     void "test creating a controller and running the test for #language and #testFramework and #buildTool"(Language language,
@@ -43,12 +35,7 @@ class CreateControllerSpec extends CommandSpec {
         1 * consoleOutput.out({ it.contains("Rendered test") })
 
         when:
-        String output = null
-        if (buildTool == BuildTool.GRADLE) {
-            output = executeGradle("test").getOutput()
-        } else if (buildTool == BuildTool.MAVEN) {
-            output = executeMaven("test")
-        }
+        String output = executeBuild(buildTool, "test")
 
         then:
         output?.contains("BUILD SUCCESS")
@@ -62,19 +49,5 @@ class CreateControllerSpec extends CommandSpec {
         "test-createcontroller-createcontrollergroovygradlejunitspec"
     }
 
-    BuildResult executeGradle(String command) {
-        BuildResult result =
-                gradleRunner.withProjectDir(dir)
-                            .withArguments(command)
-                            .build()
-        return result
-    }
 
-    String executeMaven(String command) {
-        System.setProperty("maven.multiModuleProjectDirectory", dir.absolutePath)
-        def bytesOut = new ByteArrayOutputStream()
-        PrintStream output = new PrintStream(bytesOut)
-        cli.doMain(command.split(' '), dir.absolutePath, output, output)
-        return new String(bytesOut.toByteArray(), StandardCharsets.UTF_8)
-    }
 }
