+++ test-core/src/test/groovy/io/micronaut/starter/core/test/create/CreateAppSpec.groovy
@@ -14,14 +14,10 @@ class CreateAppSpec extends CommandSpec {
         generateProject(lang, buildTool, [])
 
         when:
-        if (buildTool == BuildTool.GRADLE) {
-            executeGradleCommand('run')
-        } else {
-            executeMavenCommand("mn:run")
-        }
+        String output = executeBuild(buildTool, buildTool == BuildTool.GRADLE ? "run" : "mn:run")
 
         then:
-        testOutputContains("Startup completed")
+        output.contains("Startup completed")
 
         where:
         [lang, buildTool] << LanguageBuildCombinations.combinations()
