+++ test-utils/build.gradle
@@ -1,7 +1,26 @@
+configurations.all {
+    resolutionStrategy {
+        eachDependency { DependencyResolveDetails details ->
+            //multiple actions can be specified
+            if (details.requested.group == "org.codehaus.groovy") {
+                details.useVersion("2.5.11")
+            }
+            if (details.requested.name == "spock-core") {
+                details.useVersion("1.3-groovy-2.5")
+            }
+        }
+    }
+}
 dependencies {
     implementation("org.codehaus.groovy:groovy:$groovyVersion")
     implementation("org.spockframework:spock-core:$spockVersion") {
         exclude group: "org.codehaus.groovy", module: "groovy-all"
     }
     implementation project(":starter-core")
+    implementation gradleTestKit()
+    implementation 'org.apache.maven:maven-embedder:3.0.5'
+    implementation 'org.sonatype.aether:aether-connector-wagon:1.13.1', {
+        exclude group:'org.apache.maven.wagon', module:'wagon-provider-api'
+    }
+    implementation 'org.apache.maven.wagon:wagon-http:2.5'
 }
